seperate image gallery components into files
