+++ src/pages/components/image gallery/ImageGallery.jsx
@@ -6,67 +6,7 @@ import "./image-gallery.css";
 import "../modal.css";
 
 import { SMALL_SCREEN } from "../../../Constants";
-
-function ImageItem(props) {
-  // for loading an image when it scrolls into view
-
-  const [isVisible, setVisible] = React.useState(false);
-  const domRef = React.useRef();
-  React.useEffect(() => {
-    const observer = new IntersectionObserver((entries) => {
-      entries.forEach((entry) => setVisible(entry.isIntersecting));
-    });
-    observer.observe(domRef.current);
-    const current = domRef.current;
-    return () => observer.unobserve(current);
-  }, [isVisible]);
-
-  return (
-    <div className={`image-gallery-image  ${isVisible ? "is-visible" : ""} `}>
-      <img
-        ref={domRef}
-        alt={props.name}
-        src={props.image}
-        className="image-gallery-animation"
-        style={{
-          WebkitAnimationDelay: props.animationDelay,
-          MozAnimationDelay: props.animationDelay,
-          OAnimationDelay: props.animationDelay,
-          MsAnimationDelay: props.animationDelay,
-          AnimationDelay: props.animationDelay,
-          width: "100%",
-        }}
-      />
-    </div>
-  );
-}
-
-function ImageGalleryCollumn(props) {
-  return (
-    <>
-      {props.images.map((imageIndex) => {
-        const image = imageIndex[0];
-        const animationDelay = imageIndex[1] * 0.15 + "s";
-        const path = image.split("/");
-        const name = path[path.length - 1].split(".")[0];
-        console.log(image)
-        return (
-          <button
-            onClick={() => props.openModal(image, name)}
-            key={name}
-            className="modal-thumbnail"
-          >
-            <ImageItem
-              name={name}
-              image={image}
-              animationDelay={animationDelay}
-            />
-          </button>
-        );
-      })}
-    </>
-  );
-}
+import { ImageGalleryCollumn } from "./ImageGalleryCollumn";
 
 function ImageGallery(props) {
   const [image, setImage] = useState(null);
